PlayerWeight Fix.
Fixed a major bug with PlayerWeight. It works well for the first ~100
games. Unfortunately, the learning algorithm still needs to be tweaked.
It is learning complete nonsense right now, so the weights start to
drift after ~100 games.
Note: if testing PlayerWeight, refresh the weights using Misc.txt and
set the first line to 1.